Extract ResolutionOutputs
Configuration currently implements the 'tail-end' of resolution -- meaning it implements the process
of converting the outputs of resolution into user-consumable types like FileCollection, ArtifactCollection,
ResolvedComponentResult, etc.

We would like to reuse this implementation elsewhere to provide alternative resolution APIs.

This commit extracts a ResolutionOutputs interface which exposes the results of resolution. The intention
is for this interface to eventually become public in some form.