Deprecate ArtifactView#getAttributes
This method was left around from the time when attributes were declared directly on the artifact view instead of a
separate configurable details.

There are few if any valid use cases for needing this method, as the attributes for an artifact view are mostly
write-only. All integ tests that used artifact view did not use this method, other than a single test that specifically
tested this method.

We want to remove this method since it is difficult to implement in the new ResolutionOutputs pattern. Without this method,
ResolutionOutputs can be entirely implemented using the _outputs_ of resolution. This method requires an _input_ to resolution.

For now, we update the implementation of this method to perform resolution, consuming the input attributes from the output of resolution,
and while this works, we deprecate the method since the implementation is inefficient.