Split COnfigurationBoundExternalDependencyMetadata
This class was for dependency metadata implementations that change behavior relative to some source configuration.
The class was used to implement both Ivy and Maven dependencies, and had a flag 'alwaysUseAttributeMatching'.

For maven dependencies, this was always true. As a result, the maven instances were never actually configuration-bound anymore.

This commit splits up the configuration bound metadata into two implementations, IvyDependencyMetadata, which remains configuration-bound,
and MavenDependencyMetadata which no longer relies on a source configuration.

Eventually, we should try to merge MavenDependencyMetadata with GradleDependencyMetadata, as GradleDependenyMetadata is seemingly more
like DefaultDependencyMetadata.